Method- Harshard Number
Program to check whether a number is a Harshad number or not (Niven number) is discussed here. Harshad Number is an integer that is divisible by the sum of its digits.

 Input format:

Input consists of 1 integer.
If the given number is Harshad Number display Harshad Number or display Not Harshad Number.